seperate file for model run
