+++ tests/test_case_0dparcel/test_python_microphys.py
@@ -66,7 +66,7 @@ def test_python_microphys():
 
   ### time parameters
   time_init = 0.0 # [s]
-  time_end = 120.0 # [s]
+  time_end = 540.0 # [s]
   timestep = 1.0 # [s]
 
   ### initial thermodynamic conditions
